+++ include/sista/coordinates.cpp
@@ -14,4 +14,17 @@ namespace sista {
     Coordinates Coordinates::operator+(const Coordinates& other) const {
         return Coordinates(y + other.y, x + other.x);
     }
+    Coordinates Coordinates::operator-(const Coordinates& other) const {
+        return Coordinates(y - other.y, x - other.x);
+    }
+    Coordinates Coordinates::operator+=(const Coordinates& other) {
+        y += other.y;
+        x += other.x;
+        return *this;
+    }
+    Coordinates Coordinates::operator-=(const Coordinates& other) {
+        y -= other.y;
+        x -= other.x;
+        return *this;
+    }
 };