+++ include/sista/coordinates.hpp
@@ -17,5 +17,8 @@ namespace sista {
         bool operator==(const Coordinates&) const;
         bool operator!=(const Coordinates&) const;
         Coordinates operator+(const Coordinates&) const;
+        Coordinates operator-(const Coordinates&) const;
+        Coordinates operator+=(const Coordinates&);
+        Coordinates operator-=(const Coordinates&);
     }; // field[y][x] - y is the row, x is the column
 };